+++ NEWS.md
@@ -1,6 +1,7 @@
 # dv.edish 1.3.0-9000
 
-- For each subject, plot peak TBILI against peak ALT/AST values (divided by ULN), rather than value pairs at each visit.
+- For each subject, plot normalized peak TBILI against normalized peak ALT/AST values, with the option to plot
+  aminotransferase values for each visit.
 - Apply log value to axes.
 - Show the time between peaks.
 - Show the sequence (peak ALT/AST on or before peak TBILI most concerning).
@@ -13,7 +14,8 @@
 # dv.edish 1.3.0
 
 - Add jumping feature.
-- Arm default values can be arbitrarily specified as long as they are character strings. This supports including multiple studies with different arm values.
+- Arm default values can be arbitrarily specified as long as they are character strings. This supports including
+  multiple studies with different arm values.
 
 # dv.edish 1.2.0
 
@@ -37,7 +39,8 @@
 
 - The module is now able to deal with multiple values per subject, visit, and lab test by using the maximum value.
 - When there is no data to display, e.g., due to filter settings, the module will provide a meaningful output.
-- The module has been updated to show commonly used default values for reference lines. For the x-axis, the default value is 3, and for the y-axis, the default value is 2.
+- The module has been updated to show commonly used default values for reference lines. For the x-axis, the default
+  value is 3, and for the y-axis, the default value is 2.
 
 # dv.edish 1.0.1